chore: remove anyerror and unreachable Result<T> returns
First, this removes the anyerror dependency and instead uses thiserror
to expose parsing errors. Secondly, this makes many getx() and setx()
functions return an Option<T> instead of a Result<T, FooErr>. The main
reason for this change is that it follows the std's convention
(e.g. Vec<T>::get). Finally, this patch removes all Result<T> return
types that do not ever return an error.